refactor : request class 분리
