+++ src/main/java/controller/Controller.java
@@ -1,7 +1,7 @@
 package controller;
 
-import http.HttpRequest;
 import http.HttpResponse;
+import http.request.HttpRequest;
 
 public interface Controller {
     void service(HttpRequest request, HttpResponse response);
