+++ src/main/java/controller/UserCreateController.java
@@ -1,13 +1,13 @@
 package controller;
 
 import db.DataBase;
-import http.HttpRequest;
 import http.HttpResponse;
+import http.request.HttpRequest;
+import http.response.RedirectView;
 import model.User;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import utils.QueryStringUtils;
-import view.RedirectView;
 
 public class UserCreateController extends AbstractController {
     private static final Logger log = LoggerFactory.getLogger(UserCreateController.class);
