+++ src/main/java/http/response/View.java
@@ -1,4 +1,4 @@
-package view;
+package http.response;
 
 import http.HTTP;
 import http.HttpStatus;
