+++ src/test/java/view/RedirectViewTest.java
@@ -1,5 +1,7 @@
 package view;
 
+import http.response.RedirectView;
+import http.response.View;
 import org.junit.jupiter.api.Test;
 
 import java.io.ByteArrayOutputStream;
